pyproject.toml
[project]
name = "coser-fastapi"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "bcrypt"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff>=0.15.13",
]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
extend-select = ["I", "UP"]

[tool.ruff.lint.per-file-ignores]
"alembic/env.py" = ["E402"]
